+++ application/controllers/streamdata.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	api "app/api/bitflyer"
 	"app/config"
-	"app/infrastructure/databases"
+	"app/domain/service"
 	"log"
 	"os"
 )
@@ -15,7 +15,7 @@ func StreamIngestionData() {
 	for ticker := range tickerChannl {
 		log.Printf("action=StreamIngestionData, %v", ticker)
 		for _, duration := range config.Config.Durations {
-			isCreated := databases.CreateCandleWithDuration(ticker, ticker.ProductCode, duration)
+			isCreated := service.CreateCandleWithDuration(ticker, ticker.ProductCode, duration)
 			if isCreated == true && duration == config.Config.TradeDuration {
 				// TODO
 			}
